back/requirements.txt
Flask==2.3.3
Flask-Cors==3.1.1
Flask-Mail==0.9.1
pdfplumber==0.10.1
python-docx==0.8.11
rapidfuzz==2.15.0
